requirements/dev.txt
-r base.txt

# Testing
pytest==8.0.2
pytest-django==4.8.0
pytest-cov==4.1.0

# Code quality
black==24.2.0
flake8==7.0.0
isort==5.13.2

# Development
django-debug-toolbar==4.3.0
